refactor: split large test files along API namespace boundaries
Break 5 large test files (1,370-1,886 lines each) into 23 smaller,
navigable files organized by API namespace. No test logic changes —
pure file moves with fixture helpers extracted to *_testdata_test.go.
